requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
psycopg[binary]>=3.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
redis>=5.0.0
httpx>=0.27.0
tenacity>=8.3.0
pandas>=2.2.0
pyarrow>=16.0.0
duckdb>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
geopandas>=0.14.0
shapely>=2.0.0
python-dateutil>=2.9.0
